requirements.txt
# Core Libraries
numpy
opencv-python-headless
Pillow
streamlit

# Computer Vision & Image Processing
firebase-admin
scipy
scikit-learn

# Streamlit Components
streamlit-drawable-canvas

# Data Processing
pandas

# File & I/O Handling
requests
io
uuid
tempfile

# Optional but recommended
python-dotenv  # For environment variables management